- #110: Refactored territories into folder.
