+++ src/app/hanzhong/data/index.ts
@@ -1,2 +1,1 @@
-export { HANZHONG_TERRITORIES } from './hanzhong-territories';
 export { HANZHONG__WAR_TIERS__TECHS__BY_LEVEL, getTechById, getTechByName } from './hanzhong-techs';
