+++ src/app/hanzhong/territories/TerritoryLevelType.ts
@@ -1,6 +1,5 @@
 import type { BaseIDType } from '../../types';
-
-import type { HanzhongTerritoryResourceType } from './hanzhong-territory-resource-type';
+import type { HanzhongTerritoryResourceType } from './TerritoryResourceType';
 
 export type HanzhongTerritoryLevelType = BaseIDType & {
   tacticalPoints: number;
