+++ src/app/hanzhong/territories/index.ts
@@ -1,0 +1,4 @@
+export { HanzhongTerritories } from './Territories';
+export { HANZHONG_TERRITORIES } from './territories--data';
+export type { HanzhongTerritoryLevelType } from './TerritoryLevelType';
+export type { HanzhongTerritoryResourceType } from './TerritoryResourceType';
