+++ src/app/hanzhong/types/index.ts
@@ -6,9 +6,6 @@ export type { HanzhongInfoDataUpdaterType } from './hanzhong-info-data-updater-t
 export type { HanzhongTechLevelType } from './hanzhong-tech-level-type';
 export type { HanzhongTechResourceType } from './hanzhong-tech-resource-type';
 export type { HanzhongTechType } from './hanzhong-tech-type';
-export type { HanzhongTerritoriesType } from './hanzhong-territories-type';
-export type { HanzhongTerritoryLevelType } from './hanzhong-territory-level-type';
-export type { HanzhongTerritoryResourceType } from './hanzhong-territory-resource-type';
 export type { HanzhongWarTierTaskType } from './hanzhong-war-tier-task-type';
 export type { HanzhongWarTierType } from './hanzhong-war-tier-type';
 export type { HanzhongUserDataType } from './UserData';
