+++ src/app/hanzhong/utils/initialize-earnings.ts
@@ -1,14 +1,7 @@
 import { calculateEarningsFromCities } from '../cities';
 import { HANZHONG_TECH_IDS } from '../constants/items-ids';
-import { HANZHONG_TERRITORIES } from '../data';
-import type {
-  HanzhongBonusType,
-  HanzhongTechType,
-  HanzhongTerritoryLevelType,
-  HanzhongTerritoryResourceType,
-  HanzhongUserDataType,
-  HanzhongWarTierType,
-} from '../types';
+import { HANZHONG_TERRITORIES, type HanzhongTerritoryLevelType, type HanzhongTerritoryResourceType } from '../territories';
+import type { HanzhongBonusType, HanzhongTechType, HanzhongUserDataType, HanzhongWarTierType } from '../types';
 import { addHanzhongBonuses } from '../utils';
 import { HANZHONG_WAR_TIERS } from '../war-tiers';
